requirements.txt
keyboard
openai
python-dotenv
pyaudio
wave
playsound